+++ apps/simpleclient_windows.c
@@ -196,9 +196,9 @@ discovery(const char *anchor, const char *uri, oc_string_array_t types,
   (void)interfaces;
   (void)bm;
   int i;
-  int uri_len = strlen(uri);
+  size_t uri_len = strlen(uri);
   uri_len = (uri_len >= MAX_URI_LENGTH) ? MAX_URI_LENGTH - 1 : uri_len;
-  PRINT("\n\nDISCOVERYCB %s %s %d\n\n", anchor, uri,
+  PRINT("\n\nDISCOVERYCB %s %s %zd\n\n", anchor, uri,
         oc_string_array_get_allocated_size(types));
   for (i = 0; i < (int)oc_string_array_get_allocated_size(types); i++) {
     char *t = oc_string_array_get_item(types, i);
