+++ security/oc_store.c
@@ -31,15 +31,17 @@
 
 #define SVR_TAG_MAX (32)
 static void
-gen_svr_tag(const char *name, int device_index, char *svr_tag)
+gen_svr_tag(const char *name, size_t device_index, char *svr_tag)
 {
-  int svr_tag_len = snprintf(svr_tag, SVR_TAG_MAX, "%s_%d", name, device_index);
-  svr_tag_len = (svr_tag_len < SVR_TAG_MAX - 1) ? svr_tag_len + 1 : SVR_TAG_MAX - 1;
+  int svr_tag_len =
+    snprintf(svr_tag, SVR_TAG_MAX, "%s_%zd", name, device_index);
+  svr_tag_len =
+    (svr_tag_len < SVR_TAG_MAX - 1) ? svr_tag_len + 1 : SVR_TAG_MAX - 1;
   svr_tag[svr_tag_len] = '\0';
 }
 
 void
-oc_sec_load_doxm(int device)
+oc_sec_load_doxm(size_t device)
 {
   long ret = 0;
   oc_rep_t *rep;
@@ -84,7 +86,7 @@ oc_sec_load_doxm(int device)
 }
 
 void
-oc_sec_load_pstat(int device)
+oc_sec_load_pstat(size_t device)
 {
   long ret = 0;
   oc_rep_t *rep = 0;
@@ -130,7 +132,7 @@ oc_sec_load_pstat(int device)
 }
 
 void
-oc_sec_load_cred(int device)
+oc_sec_load_cred(size_t device)
 {
   long ret = 0;
   oc_rep_t *rep;
@@ -173,7 +175,7 @@ oc_sec_load_cred(int device)
 }
 
 void
-oc_sec_load_acl(int device)
+oc_sec_load_acl(size_t device)
 {
   long ret = 0;
   oc_rep_t *rep;
@@ -214,7 +216,7 @@ oc_sec_load_acl(int device)
 }
 
 void
-oc_sec_dump_pstat(int device)
+oc_sec_dump_pstat(size_t device)
 {
 #ifdef OC_DYNAMIC_ALLOCATION
   uint8_t *buf = oc_mem_malloc(OC_MAX_APP_DATA_SIZE);
@@ -240,7 +242,7 @@ oc_sec_dump_pstat(int device)
 }
 
 void
-oc_sec_dump_cred(int device)
+oc_sec_dump_cred(size_t device)
 {
 #ifdef OC_DYNAMIC_ALLOCATION
   uint8_t *buf = oc_mem_malloc(OC_MAX_APP_DATA_SIZE);
@@ -266,7 +268,7 @@ oc_sec_dump_cred(int device)
 }
 
 void
-oc_sec_dump_doxm(int device)
+oc_sec_dump_doxm(size_t device)
 {
 #ifdef OC_DYNAMIC_ALLOCATION
   uint8_t *buf = oc_mem_malloc(OC_MAX_APP_DATA_SIZE);
@@ -293,7 +295,7 @@ oc_sec_dump_doxm(int device)
 }
 
 void
-oc_sec_dump_acl(int device)
+oc_sec_dump_acl(size_t device)
 {
 #ifdef OC_DYNAMIC_ALLOCATION
   uint8_t *buf = oc_mem_malloc(OC_MAX_APP_DATA_SIZE);
@@ -319,7 +321,7 @@ oc_sec_dump_acl(int device)
 }
 
 void
-oc_sec_load_unique_ids(int device)
+oc_sec_load_unique_ids(size_t device)
 {
   long ret = 0;
   oc_rep_t *rep;
@@ -378,7 +380,7 @@ oc_sec_load_unique_ids(int device)
 }
 
 void
-oc_sec_dump_unique_ids(int device)
+oc_sec_dump_unique_ids(size_t device)
 {
 #ifdef OC_DYNAMIC_ALLOCATION
   uint8_t *buf = oc_mem_malloc(OC_MAX_APP_DATA_SIZE);
